small refactoring
Release note (<category, see below>): <what> <show> <why>

+++ pkg/sql/catalog/typedesc/table_implicit_record_type.go
@@ -180,7 +180,7 @@ func (v TableImplicitRecordType) DescriptorProto() *descpb.Descriptor {
 
 // ByteSize implements the Descriptor interface.
 func (v TableImplicitRecordType) ByteSize() int64 {
-	mem := int64(v.desc.ByteSize())
+	mem := v.desc.ByteSize()
 	if v.typ != nil {
 		mem += int64(v.typ.Size())
 	}
